pyproject.toml
[project]
name = "immich-classify"
version = "0.1.0"
description = "A CLI tool to classify Immich photos using Vision Language Models"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "tabulate>=0.9",
    "loguru>=0.7",
    "python-dotenv>=1.0",
]

[project.scripts]
immich-classify = "immich_classify.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/immich_classify"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
